+++ src/cadit/ifc/ifc_py_wrap.h
@@ -1,0 +1,14 @@
+//
+// Created by Kristoffer on 24.01.2025.
+//
+
+#ifndef IFC_PY_WRAP_H
+#define IFC_PY_WRAP_H
+
+#include "../../binding_core.h"
+
+
+void ifc_module(nb::module_ &m);
+
+
+#endif //IFC_PY_WRAP_H
