+++ src/cadit/ifc/ifcop.cpp
@@ -1,9 +1,6 @@
 #include "ifcop.h"
 #include "ifcparse/Ifc4x1.h"
 #include "ifcparse/IfcFile.h"
-//#include "ifcparse/IfcHierarchyHelper.h"
-//#include "ifcparse/IfcBaseClass.h"
-//#include "ifcgeom_schema_agnostic/Serialization.h"
 #define IfcSchema Ifc4x1
 
 
@@ -24,7 +21,7 @@ int read_ifc_file(const std::string &file_name) {
     std::cout << "Found " << elements->size() << " elements in " << file_name << ":" << std::endl;
 
     std::ostringstream oss;
-    for (auto element : *elements) {
+    for (const auto element : *elements) {
         element->data().toString(oss);
         oss << "\n";
         std::cout << oss.str();
@@ -33,9 +30,4 @@ int read_ifc_file(const std::string &file_name) {
     }
 
     return 0;
-}
-
-
-void ifc_module(nb::module_ &m) {
-    m.def("read_ifc_file", &read_ifc_file, "file_name"_a, "Read an ifc file");
 }
