+++ src/cadit/occt/colors.h
@@ -3,8 +3,10 @@
 #ifndef SETINSTANCECOLOR_HPP
 #define SETINSTANCECOLOR_HPP
 
-#include "../../binding_core.h"
-
-void occt_color_module(nb::module_ &m);
-
+#include <TDF_Label.hxx>
+#include <TopoDS_Shape.hxx>
+#include <Quantity_Color.hxx>
+#include <XCAFDoc_ColorTool.hxx>
+void setInstanceColorIfAvailable(XCAFDoc_ColorTool* color_tool, const TDF_Label& lab, const TopoDS_Shape& shape,
+                                 Quantity_Color& c);
 #endif // SETINSTANCECOLOR_HPP
