--- src/cadit/occt/gltf_writer.h
@@ -1,14 +1,0 @@
-//
-// Created by Kristoffer on 07/05/2023.
-//
-
-#ifndef NANO_OCCT_GLTF_WRITER_H
-#define NANO_OCCT_GLTF_WRITER_H
-
-
-class gltf_writer {
-
-};
-
-
-#endif //NANO_OCCT_GLTF_WRITER_H
