--- src/cadit/occt/occt_convert.h
@@ -1,11 +1,0 @@
-//
-// Created by ofskrand on 24.01.2024.
-//
-
-#ifndef OCCT_CONVERT_H
-#define OCCT_CONVERT_H
-#include "../../binding_core.h"
-
-void step_to_glb_module(nb::module_ &m);
-
-#endif //OCCT_CONVERT_H
