+++ src/cadit/occt/occt_py_wrap.h
@@ -1,0 +1,16 @@
+//
+// Created by Kristoffer on 24.01.2025.
+//
+
+#ifndef OCCT_PY_WRAP_H
+#define OCCT_PY_WRAP_H
+
+#include "../../binding_core.h"
+
+void occt_color_module(nb::module_ &m);
+
+void step_to_glb_module(nb::module_ &m);
+
+void step_writer_module(nb::module_ &m);
+
+#endif //OCCT_PY_WRAP_H
