+++ src/cadit/tinygltf/tiny_helpers.cpp
@@ -1,5 +1,8 @@
 #include <iostream>
 #include "../../geom/Mesh.h"
+
+#include "../../visit/tess_helpers.h"
+#include "../../helpers/helpers.h"
 #include "tiny_gltf.h"
 #include "tiny_helpers.h"
 
@@ -106,7 +109,7 @@ void AddMesh(tinygltf::Model &model, const std::string &name, const Mesh &my_mes
     material.alphaMode = "OPAQUE";
 
     // Add the material to the model
-    int materialIndex = model.materials.size();
+    const int materialIndex = model.materials.size();
     model.materials.push_back(material);
 
     // Assign the material to the primitive
@@ -121,4 +124,57 @@ void AddMesh(tinygltf::Model &model, const std::string &name, const Mesh &my_mes
 
     // Add the node to the scene
     model.scenes[0].nodes.push_back(model.nodes.size() - 1);
+}
+
+
+int write_to_gltf(const std::string& filename, const Mesh& mesh) {
+    tinygltf::Model model;
+
+    // Create a scene
+    tinygltf::Scene scene;
+    model.scenes.push_back(scene);
+    model.defaultScene = 0;
+    AddMesh(model, "mesh", mesh);
+
+    // Save to file
+    if (tinygltf::TinyGLTF gltf; !gltf.WriteGltfSceneToFile(&model, filename, true, true, true, false)) {
+        std::cerr << "Failed to write glTF file" << std::endl;
+        return -1;
+    }
+
+    return 0;
+}
+
+// take in a list of box dimensions and origins and write to step file using the AdaCPPStepWriter class
+int write_boxes_to_gltf(const std::string &filename, const std::vector<std::vector<float>> &box_origins,
+                         const std::vector<std::vector<float>> &box_dims) {
+    tinygltf::Model model;
+    std::cout << "Exporting to " << filename <<"\n";
+
+    // Create a scene
+    tinygltf::Scene scene;
+    model.scenes.push_back(scene);
+    model.defaultScene = 0;
+
+    for (int i = 0; i < box_origins.size(); i++) {
+        const auto& origin = box_origins[i];
+        const auto& dim = box_dims[i];
+        TopoDS_Solid box = create_box(origin, dim);
+        Mesh mesh = tessellate_shape(0, box, true, 1.0, false);
+        mesh.color = random_color();
+
+        std::cout << "Adding mesh" << i << "\n";
+        AddMesh(model, "mesh", mesh);
+    }
+    // If filename contains .glb set variable "glb" to true
+    bool glb = filename.find(".glb") != std::string::npos;
+
+    // Save to file
+    tinygltf::TinyGLTF gltf;
+    if (!gltf.WriteGltfSceneToFile(&model, filename, true, true, true, glb)) {
+        std::cerr << "Failed to write glTF file" << std::endl;
+        return -1;
+    }
+
+    return 0;
 }
