+++ src/fem/fem_py_wrap.h
@@ -1,0 +1,14 @@
+//
+// Created by Kristoffer on 24.01.2025.
+//
+
+#ifndef FEM_PY_WRAP_H
+#define FEM_PY_WRAP_H
+#include "../binding_core.h"
+
+
+void gmsh_module(nb::module_ &m);
+
+
+
+#endif //FEM_PY_WRAP_H
