+++ src/fem/simple_mesh.h
@@ -5,13 +5,11 @@
 
 #ifndef ADA_CPP_SIMPLE_MESH_H
 #define ADA_CPP_SIMPLE_MESH_H
-
-#include "../binding_core.h"
+#include <string>
 
 
 void simple_gmesh(const std::string& filename);
 
-void gmsh_module(nb::module_ &m);
 
 
 #endif //ADA_CPP_SIMPLE_MESH_H
