+++ src/geom/Models.cpp
@@ -9,8 +9,6 @@
 #include "GroupReference.h"
 #include "Color.h"
 
-Color::Color(float r, float g, float b, float a) : r(r), g(g), b(b), a(a) {}
-
 MeshType from_int(int value) {
     if (value < 0 || value > 6) {
         throw std::out_of_range("Invalid MeshType value");
@@ -18,9 +16,9 @@ MeshType from_int(int value) {
     return static_cast<MeshType>(value);
 }
 
-Mesh::Mesh(int id, std::vector<float> positions, std::vector<uint32_t> faces,
+Mesh::Mesh(const int id, std::vector<float> positions, std::vector<uint32_t> faces,
            std::vector<uint32_t> edges, std::vector<float> normals,
-           MeshType mesh_type, Color color, std::vector<GroupReference> group_reference)
+           const MeshType mesh_type, Color color, std::vector<GroupReference> group_reference)
         : id(id),
           positions(std::move(positions)),
           indices(std::move(faces)),
@@ -31,13 +29,13 @@ Mesh::Mesh(int id, std::vector<float> positions, std::vector<uint32_t> faces,
           group_reference(std::move(group_reference)) {}
 
 OccShape::OccShape(TopoDS_Shape shape,
-                   Color color,
-                   int num_tot_entities,
+                   const Color color,
+                   const int num_tot_entities,
                    std::optional<std::string> name)
         : shape(std::move(shape)),
           color(color),
           num_tot_entities(num_tot_entities),
           name(std::move(name)) {}
 
-GroupReference::GroupReference(int node_id, int start, int length)
+GroupReference::GroupReference(const int node_id, const int start, const int length)
         : node_id(node_id), start(start), length(length) {}
