+++ src/geom/geom_py_wrap.h
@@ -1,0 +1,9 @@
+#ifndef GEOM_PY_WRAP_H
+#define GEOM_PY_WRAP_H
+
+#include "../binding_core.h"
+
+void shape_module(nb::module_ &m);
+
+
+#endif //GEOM_PY_WRAP_H
